update to use component/enumerable
