+++ component.json
@@ -5,8 +5,9 @@
   "version": "0.2.1",
   "keywords": [],
   "dependencies": {
-    "component/emitter" : "*",
-    "component/to-function" : "*"
+    "component/emitter": "*",
+    "component/to-function": "*",
+    "component/enumerable": "0.3.1"
   },
   "development": {},
   "license": "MIT",
